Fix for CTFramesetterSuggestFrameSizeWithConstraints bug in iOS 4
CTFramesetterSuggestFrameSizeWithConstraints returns heights that are very slightly too small for small point sizes which sometimes results in the last line of text not being drawn or worse: if you only had one line of text, no text being drawn. See issue #52 for more information.

Thanks to @VilemKurz for verifying this.