Fix a bug with incorrect truncation when "numberOfLines = 1".
If there was too much text to fit on a single line, lineBreakMode was not being respected. Regardless of the value passed in, the TTTAttributedLabel would always display text only up to the last word that would completely fit.
The displayed text was "Starbucks, Market Street, San" instead of "Starbucks, Market Street, San Fra..."

The problem is that a CTFramesetter was always being used, which implicitly performs word wrap, whereas single lines require the use of a CTTypesetter to obtain the correct truncation.