Fixed a crash caused by loading TTTAttributedLabel from a UIStoryboard
When TTTAttributedLabel is set as the class for a UIView used in a
UIStoryboard it will be created without a textColor. This will cause a
crash when text is assigned to the label because
NSAttributedStringAttributesFromLabel() attempts to assign a nil
CGColor to the attribute dictionary. TTTAttributedLabel now defaults to
black if no textColor has been set.