Increase the hit radius for links.
This is a simple approach that searches for links both at the center of the tap and along the edges and corners of the surrounding 22x22 square.  A more complete but more expensive approach would be to find all character rects within a certain distance of the touch point and check each for a link in increasing order of distance from the touch.