Remove paragraph style from default link attributes.
Any time the attributed text is modified with link attributes, the individual attributes in `linkAttributes`, `inactiveLinkAttributes`, and `addedLinkAttributes` are added or removed.  As such, there's no need to include paragraph styling as it will be inherited from the owning attributed string and doing so would actually clobber the much more complex styling provided by `NSAttributedStringAttributesFromLabel` or the passed attributed string.