It is better to not block thread by dispatch_sync onto main queue
dispatch_sync onto the main queue blocks the current thread till the
main queue signals us and we are allowed to run. As we do not use the
result of the operation performed on main queue, it is better to just
use dispatch_async onto the main queue. This way, the thread on the
global default priority dispatch queue can go back to servicing other
enqueued blocks. Details of the mechanics are described in WWDC 2015
session 718 'Building Responsive and Efficient Apps with GCD'.