Correctly calculating the link range when added in the attributed truncation token.
The offset was previously calculated by taking the last line, and saying that the range of the link (and attribution token) basically started at (line length) - (attribution token length).  This only works for monospaced fonts.  My implementation looks at the first run of the truncated line, and uses its length as the starting position for the truncation token.